requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
qrcode==8.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
